+++ src/main/java/org/folio/inventory/dataimport/handlers/actions/UpdateItemEventHandler.java
@@ -76,7 +76,7 @@ public class UpdateItemEventHandler implements EventHandler {
   private static final String RECORD_ID_HEADER = "recordId";
   private static final String CHUNK_ID_HEADER = "chunkId";
   private static final Set<String> PROTECTED_STATUSES_FROM_UPDATE = new HashSet<>(Arrays.asList("Aged to lost", "Awaiting delivery", "Awaiting pickup", "Checked out", "Claimed returned", "Declared lost", "Paged", "Recently returned"));
-  private static final String CURRENT_RETRY_NUMBER = "CURRENT_RETRY_NUMBER";
+  static final String CURRENT_RETRY_NUMBER = "CURRENT_RETRY_NUMBER";
   private static final int MAX_RETRIES_COUNT = Integer.parseInt(System.getenv().getOrDefault("inventory.di.ol.retry.number", "1"));
   private static final String CURRENT_EVENT_TYPE_PROPERTY = "CURRENT_EVENT_TYPE";
   private static final String CURRENT_NODE_PROPERTY = "CURRENT_NODE";
@@ -416,5 +416,35 @@ public class UpdateItemEventHandler implements EventHandler {
       resultedItems.add(new JsonObject().put(ITEM_PATH_FIELD, new JsonObject(ItemUtil.mapToMappingResultRepresentation(currentItem))));
     }
     return resultedItems;
+  }
+  
+  private void getActualItemAndReInvokeCurrentHandler(Item item, ItemCollection itemCollection, Promise<Item> promise, DataImportEventPayload eventPayload) {
+    itemCollection.findById(item.getId())
+      .thenAccept(actualItem -> {
+        JsonObject itemAsJson = new JsonObject(ItemUtil.mapToMappingResultRepresentation(actualItem));
+        eventPayload.getContext().put(ITEM.value(), Json.encode(itemAsJson));
+        eventPayload.getEventsChain().remove(eventPayload.getContext().get(CURRENT_EVENT_TYPE_PROPERTY));
+        try {
+          eventPayload.setCurrentNode(ObjectMapperTool.getMapper().readValue(eventPayload.getContext().get(CURRENT_NODE_PROPERTY), ProfileSnapshotWrapper.class));
+        } catch (JsonProcessingException e) {
+          LOG.error("Cannot map from CURRENT_NODE value", e);
+        }
+        eventPayload.getContext().remove(CURRENT_EVENT_TYPE_PROPERTY);
+        eventPayload.getContext().remove(CURRENT_NODE_PROPERTY);
+        handle(eventPayload).whenComplete((res, e) -> {
+          if (e != null) {
+            promise.fail(e.getMessage());
+          } else {
+            promise.complete(item);
+          }
+        });
+      })
+      .thenAccept(v -> eventPayload.getContext().remove(CURRENT_RETRY_NUMBER))
+      .exceptionally(e -> {
+        eventPayload.getContext().remove(CURRENT_RETRY_NUMBER);
+        LOG.error(format("Cannot get actual Item by id: %s", e.getCause()));
+        promise.fail(format("Cannot get actual Item by id: %s", e.getCause()));
+        return null;
+      });
   }
 }